+++ Core/Contracts/RepositoryInterfaces.cs
@@ -1,0 +1,14 @@
+﻿using Core.Entities;
+
+namespace Core.Contracts
+{
+    public interface IDocumentInfoRepository
+    {
+        Task<IEnumerable<DocumentInfo>> GetAllDocumentsAsync(string ownerId, bool trackChanges);
+        Task<DocumentInfo?> GetDocumentAsync(string ownerId, Guid docId, bool trackChanges);
+        void CreateDocument(string ownerId, DocumentInfo documentInfo);
+        void DeleteDocument(DocumentInfo documentInfo);
+
+        Task SaveAsync();
+    }
+}