+++ dochost.client/src/App.tsx
@@ -1,11 +1,7 @@
 import './App.css';
 import {QueryClient, QueryClientProvider} from "@tanstack/react-query";
-import Dashboard from "./dashboard/Dashboard.tsx";
 import AuthProvider from "./auth/AuthContext.tsx";
-import {Navigate, Route, Routes} from "react-router-dom";
-import Login from "./auth/Login.tsx";
-import Layout from "./layout/Layout.tsx";
-import RequireAuth from "./layout/RequireAuth.tsx";
+import AppRoutes from "./layout/AppRoutes.tsx";
 
 const queryClient = new QueryClient();
 
@@ -14,26 +10,7 @@ function App() {
     return (
         <AuthProvider>
             <QueryClientProvider client={queryClient}>
-                <Dashboard/>
-
-                <Routes>
-                    <Route element={<Layout/>}>
-                        <Route path="/" element={<Navigate to="/dashboard" replace/>}/>
-                        <Route path="/login" element={<Login/>}/>
-                        <Route
-                            path="/dashboard"
-                            element={
-                                <RequireAuth>
-                                    <Dashboard/>
-                                </RequireAuth>
-                            }
-                        />
-                        <Route
-                            path="*"
-                            element={<Navigate to="/dashboard" replace/>}
-                        />
-                    </Route>
-                </Routes>
+                <AppRoutes />
             </QueryClientProvider>
         </AuthProvider>
     )