+++ dochost.client/src/layout/Layout.tsx
@@ -1,11 +1,11 @@
 ﻿import {Outlet} from "react-router-dom";
 import AuthStatus from "./AuthStatus.tsx";
 
-function Layout() {
+function Layout( { auth }) {
     // TODO: use flex column for the parent div so navigation is always displayed on top
     return (
         <div>
-            <AuthStatus/>
+            <AuthStatus auth={auth}/>
             <Outlet/>
         </div>
     );