+++ dochost.client/src/layout/RequireAuth.tsx
@@ -1,15 +1,13 @@
-﻿import {useAuth} from "../auth/AuthContext.tsx";
-import {Navigate, useLocation} from "react-router-dom";
+﻿import {Navigate, Outlet, useLocation} from "react-router-dom";
 
-function RequireAuth({children}: { children: JSX.Element }) {
-    const auth = useAuth();
+function RequireAuth({ session }) {
     const location = useLocation();
 
-    if (!auth.session) {
+    if (!session) {
         return <Navigate to="/login" state={{from: location}} replace/>;
     }
 
-    return children;
+    return <Outlet />;
 }
 
 export default RequireAuth