+++ dochost.client/CHANGELOG.md
@@ -1,10 +1,13 @@
 This file explains how Visual Studio created the project.
 
 The following tools were used to generate this project:
+
 - create-vite
 
 The following steps were used to generate this project:
-- Create react project with create-vite: `npm init --yes vite@latest dochost.client -- --template=react-ts`.
+
+- Create react project with create-vite:
+  `npm init --yes vite@latest dochost.client -- --template=react-ts`.
 - Update `vite.config.ts` to set up proxying and certs.
 - Add `@type/node` for `vite.config.js` typing.
 - Update `App` component to fetch and display weather information.